Set up forgotten signals to fix melting
Between you and me, I think I should get credit for this in the jam.
The only reason I didn't have it in earlier is because Godot didn't
support multi-select signal enabling. I had these selected earlier,
but I had to deselect in order to show the signal pane, forgetting
to enable more than one in my haste to submit on time.